+++ ssz/src/main/java/org/ethereum/beacon/ssz/type/HashCodec.java
@@ -53,7 +53,7 @@ public class HashCodec implements SSZCodec {
   }
 
   @Override
-  public long getSize(SSZField field) {
+  public int getSize(SSZField field) {
     return 32;
   }
