+++ ssz/src/main/java/org/ethereum/beacon/ssz/type/StringPrimitive.java
@@ -27,7 +27,7 @@ public class StringPrimitive implements SSZCodec {
   }
 
   @Override
-  public long getSize(SSZField field) {
+  public int getSize(SSZField field) {
     return -1;
   }
