+++ ssz/src/main/java/org/ethereum/beacon/ssz/type/UIntCodec.java
@@ -75,7 +75,7 @@ public class UIntCodec implements SSZCodec {
   }
 
   @Override
-  public long getSize(SSZField field) {
+  public int getSize(SSZField field) {
     return parseFieldType(field).size / 8;
   }
